Implementing User roles

+++ app.py
@@ -2,10 +2,12 @@ from flask import Flask, request, jsonify
 from models.user import User
 from database import db
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
+import bcrypt
+
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = 'your_secret_key'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud' 
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -27,7 +29,7 @@ def login():
     if username and password:
         user = User.query.filter_by(username=username).first()
     
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             print(current_user.is_authenticated)
             return jsonify({'message': 'Autenticação realizada com sucesso'})
@@ -41,19 +43,18 @@ def logout():
     return jsonify({'message': 'Logout realizado com sucesso'})
 
 @app.route('/user', methods=["POST"])
-@login_required
 def create_user():
     data = request.json
     username = data.get('username')
     password = data.get('password')
 
     if username and password:
-        user = User(username=username, password=password)
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt()) #método para criptografar as senhas cadastradas
+        user = User(username=username, password=hashed_password, role="user")
         db.session.add(user)
         db.session.commit()
         return jsonify({'message': 'Usuário cadastrado com sucesso'})
         
-
     return jsonify({'message': 'Dados inválidos'}), 400
 
 @app.route ('/user/<int:id_user>', methods=['GET']) #para mostrar o campo de senha, tenho que criptografar para proteger os dados do ususario
@@ -72,6 +73,9 @@ def update_user(id_user): #atualizando apenas a senha para nao gerar conflito de
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({'message': 'Você não tem permissão para realizar essa ação'}), 403
+    
     if user and data.get('password'):
         user.password = data.get('password')
         db.session.commit()
@@ -85,6 +89,9 @@ def update_user(id_user): #atualizando apenas a senha para nao gerar conflito de
 def delete_user(id_user):
     user = User.query.get(id_user)
 
+    if current_user.role != "admin":
+        return jsonify({'message': 'Você não tem permissão para realizar essa ação'}), 403
+
     if user and user.id != current_user.id: #se o usuário que eu estou deletando for diferente do meu usuário atual, execute:
         db.session.delete(user)
         db.session.commit()
@@ -92,11 +99,6 @@ def delete_user(id_user):
     
     return jsonify({'message': 'Deleção não permitida'}), 403
 
-
-
-@app.route('/hello-world', methods=['GET'])
-def hello_world():
-    return 'Hello, World!'
 
 if __name__ == '__main__':
     app.run(debug=True)
